requirements.txt
accelerate==0.34.2
bitsandbytes==0.43.3
dadaptation==3.2
diffusers==0.30.2
transformers==4.44.2
sentencepiece==0.2.0
lion-pytorch==0.2.2
lpips==0.1.4
matplotlib==3.9.2
numpy==1.26.4
openai==1.44.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
pandas==2.2.2
pillow==10.4.0
prodigyopt==1.0
pydantic==2.9.1
PyYAML==6.0.1
requests==2.32.3
safetensors==0.4.5
torch==2.4.0
tqdm==4.66.5
wandb==0.18.0
datasets==3.0.0
ftfy==6.2.3
scikit-learn==1.5.2
git+https://github.com/davidbau/baukit
anthropic==0.34.2
peft==0.12.0
xformers==0.0.27.post2
lm-eval